+++ tests/testthat/testCompartments.R
@@ -4,32 +4,32 @@ library(testthat)
 context("Test all methods that may be applied on compartments")
 
 test_that("add method can't be used to add a compartment (add an ODE instead)", {
-  model <- model_suite$nonmem$advan4_trans4
+  model <- model_suite$testing$nonmem$advan4_trans4
   expect_error(model %>% add(Compartment(index=model@compartments %>% length() + 1, name="OUTPUT2")),
                regexp="No default function is provided")
 })
 
 test_that("contains method works with compartments", {
-  model <- model_suite$nonmem$advan4_trans4
+  model <- model_suite$testing$nonmem$advan4_trans4
   expect_true(model %>% contains(Compartment(1)))
   expect_false(model %>% contains(Compartment(5)))
 })
 
 test_that("delete method can't be used to delete a compartment (at least for now)", {
-  model <- model_suite$nonmem$advan4_trans4
+  model <- model_suite$testing$nonmem$advan4_trans4
   expect_error(model %>% delete(Compartment(index=1)), regexp="No default function is provided")
 })
 
 test_that("find method may return a compartment", {
   # Example 1
-  model <- model_suite$nonmem$advan4_trans4
+  model <- model_suite$testing$nonmem$advan4_trans4
   compartment <- model %>% find(Compartment(index=2))
   
   expect_equal(compartment@index, 2)
   expect_equal(compartment@name, "CENTRAL")
   
   # Example 2
-  model <- model_suite$nonmem$advan1_trans1
+  model <- model_suite$testing$nonmem$advan1_trans1
   compartments <- model@compartments
   compartment1 <- compartments %>% find(Compartment(index=1))
   compartment2 <- compartments %>% find(Compartment(index=2))
@@ -40,7 +40,7 @@ test_that("find method may return a compartment", {
 })
 
 test_that("getCompartmentIndex method works as expected", {
-  model <- model_suite$nonmem$advan4_trans4
+  model <- model_suite$testing$nonmem$advan4_trans4
   index_depot <- model %>% getCompartmentIndex("DEPOT")
   index_central <- model %>% getCompartmentIndex("CENTRAL")
   
@@ -50,7 +50,7 @@ test_that("getCompartmentIndex method works as expected", {
 })
 
 test_that("replace method may be used to adapt the name of a compartment", {
-  model <- model_suite$nonmem$advan4_trans4 %>% add(Bioavailability(1, "0.75"))
+  model <- model_suite$testing$nonmem$advan4_trans4 %>% add(Bioavailability(1, "0.75"))
   
   # Replace compartment and ODE
   model <- model %>% replace(Compartment(1, "ABS")) %>%