+++ tests/testthat/testModelLibrary.R
@@ -5,6 +5,6 @@ context("Test model library")
 
 test_that("Templates can be loaded successfully", {
   
-  model <- model_suite$nonmem$advan1_trans1
+  model <- model_suite$testing$nonmem$advan1_trans1
   expect_true(model@parameters %>% contains(Theta(index=1, name="K")))
 })