+++ tests/testthat/testUtils.R
@@ -3,9 +3,6 @@
 # roxygen2::roxygenise()
 # setwd("C:/prj/campsismod/tests/")
 # testFolder <<- "C:/prj/campsismod/tests/testthat/"
-# reticulate::use_python("C:/PsN-5.0.0/python/python-3.7.7.amd64/python.exe", required=TRUE)
-# reticulate::py_config()
-# version <- pharmpy["__version__"]
 
 testFolder <- ""
 overwriteNonRegressionFiles <- FALSE